+++ packages/ubcli/CHANGELOG.md
@@ -6,7 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-
+ - new command line option `-su` (skipUndocumented) for `ubcli generateDoc`. If passed then undocumented API methods
+ will be excluded from documentation.
+ Example: `npx ubcli generateDoc -u admin -p admin -su`
+  
 ### Changed
 
 ### Deprecated